New list ui, geolocation

+++ src/app/modules/places-overview/place-table/place-table.component.ts
@@ -1,12 +1,14 @@
 import { Component } from '@angular/core';
 import { FirestoreDataService } from '../../../services/firestore-data.service';
-import { FormBuilder } from '@angular/forms';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CreateEditPlaceComponent } from '../create-edit-place/create-edit-place.component';
 import { Emoji } from '../../../models/emoji';
 import { Place } from '../../../models/place';
-import { faEdit, faTrashAlt } from '@fortawesome/free-regular-svg-icons';
+import { faEdit } from '@fortawesome/free-regular-svg-icons';
+import { faRemove } from '@fortawesome/free-solid-svg-icons';
+import { ConfirmModalComponent } from '../../../shared/components/confirm-modal/confirm-modal.component';
+import { LocationService } from '../../../services/location.service';
 
 @Component({
   selector: 'app-place-table',
@@ -17,14 +19,14 @@ export class PlaceTableComponent {
 
   Emoji = Emoji;
   faEdit = faEdit;
-  faTrashAlt = faTrashAlt;
+  faRemove = faRemove;
 
   constructor(
     public firestoreDataService: FirestoreDataService,
-    private fb: FormBuilder,
     private modalService: NgbModal,
     private router: Router,
-    private activatedRoute: ActivatedRoute
+    private activatedRoute: ActivatedRoute,
+    private locationService: LocationService
   ) {}
 
   createPlace(): void {
@@ -38,7 +40,13 @@ export class PlaceTableComponent {
   }
 
   deletePlace(id: string): void {
-    this.firestoreDataService.delete(id).subscribe();
+    const modalRef = this.modalService.open(ConfirmModalComponent, { centered: true });
+    (modalRef.componentInstance as ConfirmModalComponent).itemName = 'place';
+    modalRef.result.then(result => {
+      if (result) {
+        this.firestoreDataService.delete(id).subscribe();
+      }
+    });
   }
 
   goToDetails(item: Place): void {
